Centralize the parser location. Also add the --device flag.
Centralized the shark_args parser. Also added the --device flag that
specifies the device on which the inference or training is to be done.
